Start testing travis release builds
We want to start generating release artifacts for VSM2 from our
libarchive repo. We will have changes that are not yet upstream or
different packaging requirements from the defaults.

To start, we'll run the repo makerelease script which does a bunch of
test and verification on the release dist (tar.gz, zip).