Drop sparse tests so rpms can build in docker
The docker environment on Mac and Linux doesn't seem to generate the
sparse files that libarchive tests require. We set up a RPM only patch
to disable the problem tests.

It should be noted, these tests are still run in the native build that
is done in the testing steps before we try to verify RPMS. There is a
small risk that we introduce a sparse bug that only shows up when the
same code is rebuild for CentOS, but we can address that in future
changes.