Drop LIBARCHIVE.xattr pax headers
Libarchive double records the xattrs, in both the LIBARCHIVE.xattr and
SCHILY.xattr pax namespaces. This results in double entires on reading,
as those prefix are stripped and we are unable to differentiate. To make
it really difficult, the data in the LIBARCHIVE.xattr space is b64
encoded. It was easiest to have libarchive drop the LIBARCHIVE.xattr
namespace, rather than deal with that nonsense in python.

This is a temp hack, we'll add proper runtime toggles for this in the
future. We should also add tests for that behavior too.