Allow reading of PAX KW items
The first step in generic PAX keyword support is the reading of the
items. The API follows that of sparse and xattr, making upper layer
support easy.

A quick test validates via the the known path of xattrs -> PAX KW.

The README.versity contains some helpful hints on building and testing.
This should make it easier for future-us to get in here and keep
enhancing code.