Add more developer tooling
Now that we want to ship a Cython module from versity-python-libarchive,
we need a build environment. Our choice is always docker, as that allows
Mac hosts to play nicely and keeps the environments canned.

While here, update our script tools to make life nicer for the
developer. We now support the ability to make libarchive code changes
and have those built as RPMS and used for Python development.