Bump release to 3.3.2-vsm2
We need a new RPM release now that the PAX KW bits have landed.

The PATH tweak is to allow a developer to run the uploads when jfrog is
installed via homebrew or other methods.

While here, fix up a bad README merge that slipped through previous
commits. Also added in the brief bits on how to tag and cut a new RPM
version for release.